requirements.txt
# general modules
Django==4.2.15
pillow==10.4.0
djangorestframework==3.15.2
python-decouple==3.8

# third party modules
django-filter==24.3
Markdown==3.7
drf-yasg[validation]
djangorestframework_simplejwt
django-coreapi
django-cors-headers
djoser
django-ckeditor

# deployment modules
gunicorn
django-robots
psycopg2

# email third party modules
django-mail-templated

# refactoring and tests
black
flake8
pytest
pytest-django
faker

# background process & cache
celery
redis
django-redis
django-celery-beat